requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
Arpeggio==1.9.2
astroid==2.3.3
attrs==19.3.0
Babel==2.7.0
bcrypt==3.1.7
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.3
colorlog==4.0.2
cryptography==2.8
Django==2.0.3
django-celery==3.2.2
django-db-reconnect==0.0.1
dwebsocket==0.4.2
eventlet==0.22.1
filetype==1.0.5
flower==0.9.2
greenlet==0.4.15
har2case==0.3.1
httprunner==2.2.6
idna==2.6
isort==4.3.21
Jinja2==2.10.3
jsonpath==0.82
kombu==3.0.37
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
multi-key-dict==2.0.3
packaging==20.0
paramiko==2.4.1
parver==0.2.1
pbr==5.4.4
pipenv==2018.11.26
pyasn1==0.4.8
pycparser==2.19
pyjenkins==0.1.2
pylint==2.4.4
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.6
python-jenkins==1.6.0
pytz==2019.3
PyYAML==5.3
requests==2.18.4
requests-toolbelt==0.8.0
six==1.13.0
sqlparse==0.3.0
tornado==5.1.1
typed-ast==1.4.0
urllib3==1.22
virtualenv==16.7.9
virtualenv-clone==0.5.3
wrapt==1.11.2
XlsxWriter==1.3.3